+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Fix mods logger on launch not actually logging everything that was added
 - Fix Fabric dependency showing for mods even when installed on another mod platform
 - Fix issue with Quilt mods showing as updates for Fabric instances [#622]
+- Don't show Fabric dependency when QSL is installed
 
 ### Misc
 - Switch to using Log4j2 for application logging [#533] (thanks @s0cks)